refactor: standardise skill UI colours and add executor views
- Replace primary colour with onSurfaceVariant throughout skill‑related views.
- Introduce SkillExecutorView, SkillsGalleryView, and SkillsHistoryPanel for enhanced skill interactions.
- Update i18n strings for skill workspace dialog and search placeholder.
- Refactor colour usage in project, session, and navigation sections to use onSurface variants.
- Remove unused imports and suppress lint warnings.
