+++ desktop-shared/src/main/kotlin/io/askimo/ui/session/SessionsView.kt
@@ -364,7 +364,7 @@ private fun sortableHeader(
             Icon(
                 imageVector = if (direction == SortDirection.DESC) Icons.Default.ArrowDownward else Icons.Default.ArrowUpward,
                 contentDescription = null,
-                tint = MaterialTheme.colorScheme.primary,
+                tint = MaterialTheme.colorScheme.onSurfaceVariant,
                 modifier = Modifier.size(12.dp),
             )
         }
@@ -422,7 +422,7 @@ private fun sessionRow(
                 imageVector = if (session.isStarred) Icons.Default.Star else Icons.Outlined.StarOutline,
                 contentDescription = if (session.isStarred) "Unpin session" else "Pin session",
                 tint = if (session.isStarred) {
-                    MaterialTheme.colorScheme.primary
+                    MaterialTheme.colorScheme.onSurface
                 } else {
                     MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.4f)
                 },
@@ -482,7 +482,7 @@ private fun sessionRow(
                         Icon(
                             if (session.isStarred) Icons.Default.StarBorder else Icons.Default.Star,
                             contentDescription = null,
-                            tint = if (session.isStarred) MaterialTheme.colorScheme.onSurfaceVariant else MaterialTheme.colorScheme.primary,
+                            tint = MaterialTheme.colorScheme.onSurfaceVariant,
                         )
                     },
                     modifier = Modifier.pointerHoverIcon(PointerIcon.Hand),
