+++ desktop-shared/src/main/resources/i18n/messages.properties
@@ -1210,6 +1210,8 @@ skills.view.no.agent.hint=No external agent found. Install Claude Code (claude) 
 skills.view.response.error=Error
 skills.view.workdir.label=Working directory
 skills.view.workdir.change=Change
+skills.view.workdir.dialog.title=Select Working Directory
+skills.view.workdir.open=Open folder
 skills.view.context.placeholder=What would you like to do? (optional)
 skills.view.system.prompt.preview=System Prompt
 skills.view.response.title=Response
@@ -1223,6 +1225,7 @@ skills.view.system.prompt.file=file
 skills.view.system.prompt.files=files
 skills.view.system.prompt.also.included=Also included in agent context:
 skills.view.activity.log=Activity Log
+skills.view.search.placeholder=Search skills ...
 
 
 # File / folder chooser
